refactor(harmonizer): move release type helpers to harmonizer/release_types.ts
